requirements.txt
flask==3.0.3
gunicorn==21.2.0
selenium==4.21.0
webdriver-manager==4.0.1
pandas==2.2.2
openpyxl==3.1.2
pytesseract==0.3.13
pdfplumber==0.11.0
Pillow==10.3.0
